Save, Load, Exit missing

+++ Assets/Scripts/DeleteAllitemEditor.cs
@@ -9,9 +9,10 @@ public class DeleteAllitemEditor : Editor {
 	public override void OnInspectorGUI()
 	{
 		DeleteAllItem remove = (DeleteAllItem)target;
-		if(GUILayout.Button("Delete From Scene"))
-			remove.DeleteAllFromScreen();
-
+		if (Application.isPlaying == false) {
+			if (GUILayout.Button ("Delete From Scene"))
+				remove.DeleteAllFromScreen ();
+		}
 		if (Application.isPlaying == true) {
 			if (GUILayout.Button ("Delete From Backpack")) 
 				remove.DeleteAllFromBackpack ();
